requirements.txt
streamlit
sentence-transformers
numpy
pandas
faiss-cpu
google-generativeai
python-dotenv